fix(bookings): move weekly-call-counts helper out of the route file
Next route modules may only export HTTP-method fields — exporting the
helper failed the build; it now lives in lib/booking/weekly-call-counts.

Part of #1018.
